+++ ribs/schedulers/_bandit_scheduler.py
@@ -1,5 +1,6 @@
 """Provides the Bandit Scheduler."""
 import warnings
+from collections import defaultdict
 
 import numpy as np
 
@@ -305,22 +306,25 @@ class BanditScheduler:
         raise NotImplementedError("tell_dqd() is not supported by"
                                   "BanditScheduler.")
 
-    def tell(self, objective, measures):
+    def tell(self, objective, measures, **fields):
         """Returns info for solutions from :meth:`ask`.
 
         The emitters are the same with those used in the last call to
         :meth:`ask`.
 
-        .. note:: The objective batch and measures batch must be in the same
-            order as the solutions created by :meth:`ask_dqd`; i.e.
-            ``objective_batch[i]`` and ``measures_batch[i]`` should be the
-            objective and measures for ``solution_batch[i]``.
+        .. note:: The objective and measures arrays must be in the same order as
+            the solutions created by :meth:`ask_dqd`; i.e. ``objective[i]`` and
+            ``measures[i]`` should be the objective and measures for
+            ``solution[i]``.
 
         Args:
             objective_batch ((batch_size,) array): Each entry of this array
                 contains the objective function evaluation of a solution.
             measures_batch ((batch_size, measures_dm) array): Each row of
                 this array contains a solution's coordinates in measure space.
+            fields (keyword arguments): Additional data for each solution. Each
+                argument should be an array with batch_size as the first
+                dimension.
         Raises:
             RuntimeError: This method is called without first calling
                 :meth:`ask`.
@@ -333,6 +337,7 @@ class BanditScheduler:
         data = self._validate_tell_data({
             "objective": objective,
             "measures": measures,
+            **fields,
         })
 
         archive_empty_before = self.archive.empty
@@ -343,27 +348,26 @@ class BanditScheduler:
 
         # Add solutions to the archive.
         if self._add_mode == "batch":
-            status_batch, value_batch = self.archive.add(**data)
+            add_info = self.archive.add(**data)
 
             # Add solutions to result_archive.
             if self._result_archive is not None:
                 self._result_archive.add(**data)
         elif self._add_mode == "single":
-            status_batch = []
-            value_batch = []
+            add_info = defaultdict(list)
 
             for i in range(len(self._cur_solutions)):
                 single_data = {name: arr[i] for name, arr in data.items()}
-                status, value = self.archive.add_single(**single_data)
-                status_batch.append(status)
-                value_batch.append(value)
+                single_info = self.archive.add_single(**single_data)
+                for name, val in single_info.items():
+                    add_info[name].append(val)
 
                 # Add solutions to result_archive.
                 if self._result_archive is not None:
                     self._result_archive.add_single(**single_data)
 
-            status_batch = np.asarray(status_batch)
-            value_batch = np.asarray(value_batch)
+            for name, arr in add_info.items():
+                add_info[name] = np.asarray(arr)
 
         # Warn the user if nothing was inserted into the archives.
         if archive_empty_before and self.archive.empty:
@@ -381,8 +385,13 @@ class BanditScheduler:
 
             end = pos + n
             self._selection[i] += n
-            self._success[i] += np.count_nonzero(status_batch[pos:end])
-            emitter.tell(self._cur_solutions[pos:end],
-                         data["objective"][pos:end], data["measures"][pos:end],
-                         status_batch[pos:end], value_batch[pos:end])
+            self._success[i] += np.count_nonzero(add_info["status"][pos:end])
+            emitter.tell(
+                **{
+                    name: arr[pos:end] for name, arr in data.items()
+                },
+                add_info={
+                    name: arr[pos:end] for name, arr in add_info.items()
+                },
+            )
             pos = end