+++ ribs/schedulers/_scheduler.py
@@ -1,5 +1,6 @@
 """Provides the Scheduler."""
 import warnings
+from collections import defaultdict
 
 import numpy as np
 
@@ -233,27 +234,26 @@ class Scheduler:
 
         # Add solutions to the archive.
         if self._add_mode == "batch":
-            status_batch, value_batch = self.archive.add(**data)
+            add_info = self.archive.add(**data)
 
             # Add solutions to result_archive.
             if self._result_archive is not None:
                 self._result_archive.add(**data)
         elif self._add_mode == "single":
-            status_batch = []
-            value_batch = []
+            add_info = defaultdict(list)
 
             for i in range(len(self._cur_solutions)):
                 single_data = {name: arr[i] for name, arr in data.items()}
-                status, value = self.archive.add_single(**single_data)
-                status_batch.append(status)
-                value_batch.append(value)
+                single_info = self.archive.add_single(**single_data)
+                for name, val in single_info.items():
+                    add_info[name].append(val)
 
                 # Add solutions to result_archive.
                 if self._result_archive is not None:
                     self._result_archive.add_single(**single_data)
 
-            status_batch = np.asarray(status_batch)
-            value_batch = np.asarray(value_batch)
+            for name, arr in add_info.items():
+                add_info[name] = np.asarray(arr)
 
         # Warn the user if nothing was inserted into the archives.
         if archive_empty_before and self.archive.empty:
@@ -262,15 +262,15 @@ class Scheduler:
             if result_archive_empty_before and self.result_archive.empty:
                 warnings.warn(self.EMPTY_WARNING.format(name="result_archive"))
 
-        return status_batch, value_batch
+        return add_info
 
-    def tell_dqd(self, objective, measures, jacobian):
+    def tell_dqd(self, objective, measures, jacobian, **fields):
         """Returns info for solutions from :meth:`ask_dqd`.
 
-        .. note:: The objective batch, measures batch, and jacobian batch must
-            be in the same order as the solutions created by :meth:`ask_dqd`;
-            i.e. ``objective[i]``, ``measures[i]``, and ``jacobian[i]`` should
-            be the objective, measures, and jacobian for ``solution[i]``.
+        .. note:: The objective, measures, and jacobian arrays must be in the
+            same order as the solutions created by :meth:`ask_dqd`; i.e.
+            ``objective[i]``, ``measures[i]``, and ``jacobian[i]`` should be the
+            objective, measures, and jacobian for ``solution[i]``.
 
         Args:
             objective ((batch_size,) array): Each entry of this array contains
@@ -282,6 +282,9 @@ class Scheduler:
                 solutions obtained from :meth:`ask_dqd`. Each matrix should
                 consist of the objective gradient of the solution followed by
                 the measure gradients.
+            fields (keyword arguments): Additional data for each solution. Each
+                argument should be an array with batch_size as the first
+                dimension.
         Raises:
             RuntimeError: This method is called without first calling
                 :meth:`ask`.
@@ -295,36 +298,45 @@ class Scheduler:
         data = self._validate_tell_data({
             "objective": objective,
             "measures": measures,
+            **fields,
         })
 
         jacobian = np.asarray(jacobian)
         self._check_length("jacobian", jacobian)
 
-        status_batch, value_batch = self._add_to_archives(data)
+        add_info = self._add_to_archives(data)
 
         # Keep track of pos because emitters may have different batch sizes.
         pos = 0
         for emitter, n in zip(self._emitters, self._num_emitted):
             end = pos + n
-            emitter.tell_dqd(self._cur_solutions[pos:end],
-                             data["objective"][pos:end],
-                             data["measures"][pos:end], jacobian[pos:end],
-                             status_batch[pos:end], value_batch[pos:end])
+            emitter.tell_dqd(
+                **{
+                    name: arr[pos:end] for name, arr in data.items()
+                },
+                jacobian=jacobian[pos:end],
+                add_info={
+                    name: arr[pos:end] for name, arr in add_info.items()
+                },
+            )
             pos = end
 
-    def tell(self, objective, measures):
+    def tell(self, objective, measures, **fields):
         """Returns info for solutions from :meth:`ask`.
 
-        .. note:: The objective batch and measures batch must be in the same
-            order as the solutions created by :meth:`ask_dqd`; i.e.
-            ``objective[i]`` and ``measures[i]`` should be the objective and
-            measures for ``solution[i]``.
+        .. note:: The objective and measures arrays must be in the same order as
+            the solutions created by :meth:`ask_dqd`; i.e. ``objective[i]`` and
+            ``measures[i]`` should be the objective and measures for
+            ``solution[i]``.
 
         Args:
             objective ((batch_size,) array): Each entry of this array contains
                 the objective function evaluation of a solution.
             measures ((batch_size, measures_dm) array): Each row of this array
                 contains a solution's coordinates in measure space.
+            fields (keyword arguments): Additional data for each solution. Each
+                argument should be an array with batch_size as the first
+                dimension.
         Raises:
             RuntimeError: This method is called without first calling
                 :meth:`ask`.
@@ -337,15 +349,21 @@ class Scheduler:
         data = self._validate_tell_data({
             "objective": objective,
             "measures": measures,
+            **fields,
         })
 
-        status_batch, value_batch = self._add_to_archives(data)
+        add_info = self._add_to_archives(data)
 
         # Keep track of pos because emitters may have different batch sizes.
         pos = 0
         for emitter, n in zip(self._emitters, self._num_emitted):
             end = pos + n
-            emitter.tell(self._cur_solutions[pos:end],
-                         data["objective"][pos:end], data["measures"][pos:end],
-                         status_batch[pos:end], value_batch[pos:end])
+            emitter.tell(
+                **{
+                    name: arr[pos:end] for name, arr in data.items()
+                },
+                add_info={
+                    name: arr[pos:end] for name, arr in add_info.items()
+                },
+            )
             pos = end